CRUD completo de wines

+++ app/Traits/CRUDOperations.php
@@ -47,10 +47,24 @@ trait CRUDOperations
         return $model;
     }
 
+
     public function delete(Model $model): ?bool
     {
+        // Check if the model has a 'deleteChecks' method
+        if (method_exists($this, 'deleteChecks')) {
+            // If it exists, call the 'deleteChecks' method
+            $this->deleteChecks($model);
+        }
+
+       // Si las comprobaciones pasan, procedemos con la eliminación
+    $deleted = $model->delete();
+
+    // Si la eliminación fue exitosa, entonces borramos la imagen
+    if ($deleted) { //esto es para evitar errores relacionados con la imagen//
         UploadService::delete($model->image);
-        return $model->delete();
+    }
+
+    return $deleted;
     }
 
 }
